// backend/migrations/snapshots/0edc6d6c036f68de17808305b1572c961a12bd6278681104e993c50b9407f037/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Association: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly contactPersonId: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly AssociationClub: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly associationId: CodecTypes['pg/text@1']['output'];
      readonly clubId: CodecTypes['pg/text@1']['output'];
    };
    readonly Club: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly isFinancialMember: CodecTypes['pg/bool@1']['output'];
      readonly contactPersonId: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly ClubMembership: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly personId: CodecTypes['pg/text@1']['output'];
      readonly clubId: CodecTypes['pg/text@1']['output'];
      readonly isPrimaryClub: CodecTypes['pg/bool@1']['output'];
      readonly isFinancialMember: CodecTypes['pg/bool@1']['output'];
    };
    readonly Competition: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly associationId: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Match: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly sectionId: CodecTypes['pg/int4@1']['output'];
      readonly homeTeamId: CodecTypes['pg/int4@1']['output'];
      readonly awayTeamId: CodecTypes['pg/int4@1']['output'];
      readonly matchDate: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly matchStatus: CodecTypes['pg/text@1']['output'];
      readonly enteredByPersonId: CodecTypes['pg/text@1']['output'] | null;
      readonly confirmedByPersonId: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Person: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly personCode: CodecTypes['pg/text@1']['output'];
      readonly firstName: CodecTypes['pg/text@1']['output'];
      readonly lastName: CodecTypes['pg/text@1']['output'];
      readonly dateOfBirth: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly utrId: CodecTypes['pg/text@1']['output'] | null;
      readonly tennisAustraliaNumber: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly PersonRole: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly personId: CodecTypes['pg/text@1']['output'];
      readonly role: CodecTypes['pg/text@1']['output'];
      readonly clubId: CodecTypes['pg/text@1']['output'] | null;
      readonly associationId: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Rubber: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly matchId: CodecTypes['pg/text@1']['output'];
      readonly rubberType: CodecTypes['pg/text@1']['output'];
      readonly winningTeamId: CodecTypes['pg/int4@1']['output'] | null;
      readonly outcomeType: CodecTypes['pg/text@1']['output'];
    };
    readonly RubberPlayer: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly rubberId: CodecTypes['pg/int4@1']['output'];
      readonly teamId: CodecTypes['pg/int4@1']['output'];
      readonly personId: CodecTypes['pg/text@1']['output'];
    };
    readonly RubberSet: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly rubberId: CodecTypes['pg/int4@1']['output'];
      readonly setNumber: CodecTypes['pg/int4@1']['output'];
      readonly homeGames: CodecTypes['pg/int4@1']['output'];
      readonly awayGames: CodecTypes['pg/int4@1']['output'];
    };
    readonly Season: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly competitionId: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly startDate: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly endDate: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Section: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly seasonId: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Team: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly clubId: CodecTypes['pg/text@1']['output'];
      readonly sectionId: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly TeamPlayer: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly teamId: CodecTypes['pg/int4@1']['output'];
      readonly personId: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Association: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly contactPersonId: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly AssociationClub: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly associationId: CodecTypes['pg/text@1']['input'];
      readonly clubId: CodecTypes['pg/text@1']['input'];
    };
    readonly Club: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly isFinancialMember: CodecTypes['pg/bool@1']['input'];
      readonly contactPersonId: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly ClubMembership: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly personId: CodecTypes['pg/text@1']['input'];
      readonly clubId: CodecTypes['pg/text@1']['input'];
      readonly isPrimaryClub: CodecTypes['pg/bool@1']['input'];
      readonly isFinancialMember: CodecTypes['pg/bool@1']['input'];
    };
    readonly Competition: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly associationId: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Match: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly sectionId: CodecTypes['pg/int4@1']['input'];
      readonly homeTeamId: CodecTypes['pg/int4@1']['input'];
      readonly awayTeamId: CodecTypes['pg/int4@1']['input'];
      readonly matchDate: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly matchStatus: CodecTypes['pg/text@1']['input'];
      readonly enteredByPersonId: CodecTypes['pg/text@1']['input'] | null;
      readonly confirmedByPersonId: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Person: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly personCode: CodecTypes['pg/text@1']['input'];
      readonly firstName: CodecTypes['pg/text@1']['input'];
      readonly lastName: CodecTypes['pg/text@1']['input'];
      readonly dateOfBirth: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly utrId: CodecTypes['pg/text@1']['input'] | null;
      readonly tennisAustraliaNumber: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly PersonRole: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly personId: CodecTypes['pg/text@1']['input'];
      readonly role: CodecTypes['pg/text@1']['input'];
      readonly clubId: CodecTypes['pg/text@1']['input'] | null;
      readonly associationId: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Rubber: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly matchId: CodecTypes['pg/text@1']['input'];
      readonly rubberType: CodecTypes['pg/text@1']['input'];
      readonly winningTeamId: CodecTypes['pg/int4@1']['input'] | null;
      readonly outcomeType: CodecTypes['pg/text@1']['input'];
    };
    readonly RubberPlayer: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly rubberId: CodecTypes['pg/int4@1']['input'];
      readonly teamId: CodecTypes['pg/int4@1']['input'];
      readonly personId: CodecTypes['pg/text@1']['input'];
    };
    readonly RubberSet: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly rubberId: CodecTypes['pg/int4@1']['input'];
      readonly setNumber: CodecTypes['pg/int4@1']['input'];
      readonly homeGames: CodecTypes['pg/int4@1']['input'];
      readonly awayGames: CodecTypes['pg/int4@1']['input'];
    };
    readonly Season: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly competitionId: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly startDate: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly endDate: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Section: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly seasonId: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Team: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly clubId: CodecTypes['pg/text@1']['input'];
      readonly sectionId: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly TeamPlayer: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly teamId: CodecTypes['pg/int4@1']['input'];
      readonly personId: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;